+++ CHANGELOG.md
@@ -1,4 +1,4 @@
-## 0.2.4 (WIP)
+## 0.2.4 (2023-02-26)
 
 * Don't look for a field when the path is not specified
 * Make `al-field-schema` elements use `Field#getSubSchema()` method